+++ backend/libs/shared/provider/todo/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/todo.module';
+export * from './lib/todo.service';